fix: escape env variables
With this change, we avoid the problem of having to escape env variables
altogether. Previously, env variables were loaded both by bash (in development)
and docker-compose (in production). This meant that we had to deal with
escaping in both contexts, which was the source of trouble. We resolve this by
loading env vars only with bash.

Close #61.

(cherry picked from commit 9b1b08f2d9e9dfa0faec54dddd9443930ca56b12)